Migrations and models complete work along with relations in migrations

+++ national-savings-bond/database/migrations/2020_05_07_023848_create_bonds_table.php
@@ -15,7 +15,7 @@ class CreateBondsTable extends Migration
     {
         Schema::create('bonds', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('bond_type_id')->nullable();
+            $table->unsignedBigInteger('bond_type_id')->nullable();
             $table->string('serial_no')->nullable();
             $table->date('start_date')->nullable();
             $table->SoftDeletes();
